refactor: revert to original tests filenames
