+++ tsconfig.json
@@ -9,5 +9,5 @@
 		"skipLibCheck": true,
 		"strict": true
 	},
-	"include": ["*.ts", "test-*.ts", "test/many-keys-map-types.test.ts"]
+	"include": ["*.ts"]
 }
